Reorganized esbuild  files
